HR-108 Refactor code to remove duplication
